+++ src/Hasso.Cli.Tests/Units/FragmentWriterTests.Fixture.cs
@@ -1,4 +1,4 @@
-﻿namespace Hasso.Cli.Tests
+﻿namespace Hasso.Cli.Tests.Units
 {
     public class FragmentWriterFixture
     {
